requirements.txt
fastapi
uvicorn
openai
langchain
langchain_openai
langserve[all]
python-dotenv
